cambios del weonsocket

+++ src/main/java/com/sanosysalvos/notification/model/Notificacion.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Data;
 import java.time.LocalDateTime;
 
+
 @Entity
 @Table(name = "notificaciones")
 @Data
@@ -11,10 +12,10 @@ public class Notificacion {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id_notificacion;
-    private String estado_notificacion; 
+    private String estado_notificacion;
     private String descripcion;
-    private Integer id_coincidencia;
-    private Integer id_usuario;
+    private Long id_coincidencia;
+    private Long id_usuario;
     private LocalDateTime fecha_creacion = LocalDateTime.now();
     private String mensajeError;
 }
